pyproject.toml
[project]
name = "if-curator"
version = "0.1.0"
description = "Immich to Frigate training sets"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "insightface>=0.7.3",
    "numpy>=2.2.6",
    "onnxruntime>=1.23.2",
    "opencv-python>=4.12.0.88",
    "opencv-python-headless>=4.12.0.88",
    "pillow>=12.1.0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "rich>=14.2.0",
    "scipy>=1.17.0",
    "sentencepiece>=0.2.1",
    "torch>=2.9.1",
    "transformers>=4.57.6",
    "ultralytics>=8.3.252",
]

[project.optional-dependencies]
gpu = [
    "onnxruntime-gpu>=1.23.2",
]

[dependency-groups]
dev = [
    "ruff>=0.9.2",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]

